+++ CHANGELOG.md
@@ -19,7 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Changed offline emissions grid resolution templates in config files to be more descriptive
 - Read `obspack_id` from netCDF files into a character array, then convert to string
 - Add `#SBATCH -c 1` to GCHP integration test scripts and sample run scripts for Harvard Cannon
-- In GCC/GCHP integration test, passing `-s none` will run compile-only tests.  Query user to proceed or to exit.
+- In GCC/GCHP integration tests, passing `-s none` will run compile-only tests.  Query user to proceed or to exit.
+- GCC/GCHP integration tests will exit immediately if `scheduler` is omitted.
 
 ### Fixed
 - Fixed unit conversions in GEOS-only code